Refactored the issue finders (#50)
