+++ deptry/core.py
@@ -4,11 +4,11 @@ from typing import Dict, List
 
 from deptry.dependency_getter import DependencyGetter
 from deptry.import_parser import ImportParser
-from deptry.missing_dependencies_finder import MissingDependenciesFinder
+from deptry.issue_finders.missing import MissingDependenciesFinder
+from deptry.issue_finders.obsolete import ObsoleteDependenciesFinder
+from deptry.issue_finders.transitive import TransitiveDependenciesFinder
 from deptry.module import Module
-from deptry.obsolete_dependencies_finder import ObsoleteDependenciesFinder
 from deptry.python_file_finder import PythonFileFinder
-from deptry.transitive_dependencies_finder import TransitiveDependenciesFinder
 
 
 class Core:
@@ -47,21 +47,23 @@ class Core:
         all_python_files = PythonFileFinder(
             ignore_directories=self.ignore_directories, ignore_notebooks=self.ignore_notebooks
         ).get_all_python_files_in(Path("."))
+
         imported_modules = ImportParser().get_imported_modules_for_list_of_files(all_python_files)
         imported_modules = [Module(mod, dependencies) for mod in imported_modules]
+        imported_modules = [mod for mod in imported_modules if not mod.is_standard_library()]
 
         result = {}
         if not self.skip_obsolete:
             result["obsolete"] = ObsoleteDependenciesFinder(
-                imported_modules=imported_modules, dependencies=dependencies, ignore_obsolete=self.ignore_obsolete
+                imported_modules=imported_modules, dependencies=dependencies, list_to_ignore=self.ignore_obsolete
             ).find()
         if not self.skip_missing:
             result["missing"] = MissingDependenciesFinder(
-                imported_modules=imported_modules, dependencies=dependencies, ignore_missing=self.ignore_missing
+                imported_modules=imported_modules, dependencies=dependencies, list_to_ignore=self.ignore_missing
             ).find()
         if not self.skip_transitive:
             result["transitive"] = TransitiveDependenciesFinder(
-                imported_modules=imported_modules, dependencies=dependencies, ignore_transitive=self.ignore_transitive
+                imported_modules=imported_modules, dependencies=dependencies, list_to_ignore=self.ignore_transitive
             ).find()
 
         return result
